refactor host details; TS and hooks
